@@ -10,12 +10,18 @@
     "@nextui-org/system": "^2.4.6",
     "@radix-ui/react-dialog": "^1.1.4",
     "@radix-ui/react-slot": "^1.1.1",
+    "@typescript-eslint/eslint-plugin": "^8.20.0",
+    "@typescript-eslint/parser": "^8.20.0",
     "autoprefixer": "^10.4.20",
     "blakejs": "^1.2.1",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-react": "^7.37.4",
+    "eslint-plugin-unused-imports": "^4.1.4",
     "lucide-react": "^0.471.1",
-    "next": "^14.1.4",
+    "next": "^15.1.4",
     "next-themes": "^0.4.4",
     "postcss": "^8.5.0",
     "prop-types": "^15.8.1",
@@ -27,6 +33,10 @@
   },
   "devDependencies": {
     "@types/node": "22.10.6",
+    "eslint": "^9.18.0",
+    "eslint-config-prettier": "^10.0.1",
+    "eslint-plugin-prettier": "^5.2.1",
+    "prettier": "^3.4.2",
     "tailwindcss": "^3.4.17",
     "typescript": "5.7.3"
   },
