@@ -4,5 +4,5 @@ import dynamic from "next/dynamic";
 const WalletComponent = dynamic(() => import("./connector"), { ssr: false });
 
 export default function WalletConnector() {
-    return <WalletComponent />;
+  return <WalletComponent />;
 }
