@@ -18,7 +18,7 @@ const AccountLeftMenu = ({ className = "" }) => {
 
   return (
     <div className={[styles.accountLeftMenu, className].join(" ")}>
-      <div className={styles.profile} onClick={onProfileContainerClick}>
+      <div className={styles.profile} onClick={onProfileContainerClick} role="button">
         <div className={styles.profileBackground} />
         <Image
           className={styles.ioniconppersondefault}
@@ -30,7 +30,7 @@ const AccountLeftMenu = ({ className = "" }) => {
         />
         <div className={styles.profile1}>Profile</div>
       </div>
-      <div className={styles.signOut} onClick={onSignOutContainerClick}>
+      <div className={styles.signOut} onClick={onSignOutContainerClick} role="button">
         <div className={styles.profileBackground} />
         <div className={styles.signOut1}>Sign Out</div>
         <Image
