@@ -130,7 +130,8 @@ const Background = ({ className = "" }) => {
         SIGN UP
       </Button>
       <div className={styles.signinPrompt}>
-        <div className={styles.alreadyHaveAn} onClick={onAlreadyHaveAnClick}>
+        <div className={styles.alreadyHaveAn} onClick={onAlreadyHaveAnClick}
+          role="button">
           Already have an account? Sign in
         </div>
       </div>
