@@ -53,7 +53,9 @@ const NavigationMenu = ({ className = "" }) => {
         </div>
       </div>
       <div className={styles.menuOptions}>
-        <div className={styles.leftMenu2} onClick={onLeftMenuClick}>
+        <div className={styles.leftMenu2}
+          role="button"
+          onClick={onLeftMenuClick}>
           <div className={styles.homeLinkContainer}>
             <div className={styles.nestedMenuItemChild} />
             <Image
@@ -71,7 +73,9 @@ const NavigationMenu = ({ className = "" }) => {
         </div>
       </div>
       <div className={styles.menuOptions1}>
-        <div className={styles.leftMenu2} onClick={onLeftMenuClick1}>
+        <div className={styles.leftMenu2}
+          role="button"
+          onClick={onLeftMenuClick1}>
           <div className={styles.homeLinkContainer}>
             <div className={styles.nestedMenuItemChild} />
             <Image
