@@ -12,11 +12,12 @@ import {
 } from "./plutus";
 import { getPolicyId } from "@/lib/utils";
 
-export const identificationPolicyid: Data = process.env.NEXT_PUBLIC_IDENTIFICATION_PID as string;
+export const identificationPolicyid: Data = process.env
+  .NEXT_PUBLIC_IDENTIFICATION_PID as string;
 
 //------------------------------------------------------------------
 const identificationNFT_Mint = applyDoubleCborEncoding(
-  identification_nft_identification_nft_mint
+  identification_nft_identification_nft_mint,
 );
 
 export function IdentificationNFT_MintValidator(params: any[]): Validator {
@@ -28,7 +29,7 @@ export function IdentificationNFT_MintValidator(params: any[]): Validator {
 
 //------------------------------------------------------------------
 const configdatumholderscript = applyDoubleCborEncoding(
-  config_datum_holder_config_datum_holder_spend
+  config_datum_holder_config_datum_holder_spend,
 );
 
 export function ConfigDatumHolderValidator(): Validator {
@@ -40,7 +41,7 @@ export function ConfigDatumHolderValidator(): Validator {
 
 // --------------------------------------------------------------
 const ValidatorContractScript = applyDoubleCborEncoding(
-  validator_contract_validator_contract_mint
+  validator_contract_validator_contract_mint,
 );
 export function ValidatorContract(): Validator {
   //config_nft : PolicyId; validator_contract_mint: PolicyId
@@ -56,7 +57,7 @@ export function ValidatorContract(): Validator {
 
 // --------------------------------------------------------------
 const ValidatorMinterScript = applyDoubleCborEncoding(
-  validator_contract_validator_contract_mint_mint
+  validator_contract_validator_contract_mint_mint,
 );
 export function ValidatorMinter(): Validator {
   //config_nft : PolicyId;
