@@ -1,16 +1,20 @@
 /** @type {import('next').NextConfig} */
 
 const nextConfig = {
-    webpack: (config) => {
-        config.experiments = {
-            ...config.experiments,
-            asyncWebAssembly: true,
-            topLevelAwait: true,
-            layers: true,
-        };
-        return config;
-    },
-    pageExtensions: ["js", "jsx", "ts", "tsx"],
+  webpack: (config) => {
+    config.experiments = {
+      ...config.experiments,
+      asyncWebAssembly: true,
+      topLevelAwait: true,
+      layers: true,
+    };
+    config.module.rules.push({
+      test: /\.wasm$/,
+      type: "webassembly/async",
+    });
+    return config;
+  },
+  pageExtensions: ["js", "jsx", "ts", "tsx"],
 };
 
 module.exports = nextConfig;
