@@ -8,7 +8,7 @@ import { useRouter } from "next/router";
 import "./global.css";
 import WalletProvider from "@/context/walletProvider";
 
-export default function MyApp(props: { Component: any; pageProps: any; }) {
+export default function MyApp(props: { Component: any; pageProps: any }) {
   const router = useRouter();
   const { Component, pageProps } = props;
   useEffect(() => {
@@ -40,6 +40,5 @@ export default function MyApp(props: { Component: any; pageProps: any; }) {
         </ThemeProvider>
       </Fragment>
     </NextUIProvider>
-
   );
 }
